Added new DTO class to use for easier conversion to JSON and Report files, Closes #8

+++ src/main/java/com/praqma/gitrepodoctor/CreateStatistics.java
@@ -14,8 +14,13 @@ import java.util.List;
  */
 public class CreateStatistics implements CreateStatisticsIF{
 
+    /**
+     *
+     * @param files
+     * @return Long representing Actual repository size in Kilobyte
+     */
     @Override
-    public double getRepoSize(List<File> files) {
+    public Long getRepoSize(List<File> files) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
